Extract msbuild invocation method to utility class
Other tests need to invoke msbuild.exe under various execution hosts.
This functionality was extracted outside of XMake_Tests into a
RunnerUtilities class.
