+++ src/XMakeCommandLine/MSBuildTaskHost/MSBuildTaskHost/MSBuildTaskHost.csproj
@@ -93,6 +93,10 @@
     <Compile Include="..\..\..\Shared\ExceptionHandling.cs">
       <Link>ExceptionHandling.cs</Link>
     </Compile>
+    <Compile Include="..\..\..\Shared\RunnerUtilities.cs">
+      <Link>RunnerUtilities.cs</Link>
+      <ExcludeFromStyleCop>true</ExcludeFromStyleCop>
+    </Compile>
     <Compile Include="..\..\..\Shared\FileUtilities.cs">
       <Link>FileUtilities.cs</Link>
       <ExcludeFromStyleCop>true</ExcludeFromStyleCop>
